+++ packages/frontend/src/components/Routing.js
@@ -16,6 +16,7 @@ import * as accountActions from '../redux/actions/account';
 import { selectAccountSlice } from '../redux/slices/account';
 import { actions as tokenFiatValueActions } from '../redux/slices/tokenFiatValues';
 import { LoginWrapper } from '../routes/LoginWrapper';
+import { SignWrapper } from '../routes/SignWrapper';
 import translations_en from '../translations/en.global.json';
 import translations_pt from '../translations/pt.global.json';
 import translations_ru from '../translations/ru.global.json';
@@ -66,7 +67,6 @@ import Navigation from './navigation/Navigation';
 import { Profile } from './profile/Profile';
 import { ReceiveContainerWrapper } from './receive-money/ReceiveContainerWrapper';
 import { SendContainerWrapper } from './send/SendContainerWrapper';
-import { SignWrapper } from './sign/v2/SignWrapper';
 import { StakingContainer } from './staking/StakingContainer';
 import Terms from './terms/Terms';
 import { Wallet } from './wallet/Wallet';
